requirements.txt
zipfile
fastapi
uvicorn
pandas
scikit-learn
requests
datetime
calendar
re
ast
pydantic
typing 
os
sklearn.neighbors 
sklearn.compose 
sklearn.preprocessing 
sklearn.pipeline 
Pipeline
NearestNeighbors
ColumnTransformer
OneHotEncoder
StandardScaler